backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
